backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
neo4j==5.15.0

# AI/ML框架
torch==2.1.1
transformers==4.36.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
opencv-python==4.8.1.78
pillow==10.1.0

# 异步支持
asyncio==3.4.3
httpx==0.25.2
websockets==12.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 工具库
celery==5.3.4
loguru==0.7.2
rich==13.7.0
typer==0.9.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 监控
prometheus-client==0.19.0
structlog==23.2.0

# 文件处理
aiofiles==23.2.1
python-magic==0.4.27

# 图像处理
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# 知识图谱
py2neo==2021.2.4
networkx==3.2.1

# 机器学习扩展
xgboost==2.0.2
lightgbm==4.1.0
optuna==3.4.0

# 自然语言处理
spacy==3.7.2
nltk==3.8.1

# 向量数据库
chromadb==0.4.18
faiss-cpu==1.7.4

# Web3 (可选的区块链功能)
web3==6.13.0
eth-account==0.10.0 